+++ src/hooks/useFilters.tsx
@@ -1,0 +1,11 @@
+import { useParams } from 'hooks/useParams';
+
+export function useFilters() {
+  return useParams({
+    q: 'string',
+    page: {
+      type: 'number',
+      default: 0,
+    },
+  });
+}
